move project error code to api/interface as suggested
