--- cue/model/api/v1/project_error_go_gen.cue
@@ -1,8 +1,0 @@
-// Code generated by cue get go. DO NOT EDIT.
-
-//cue:generate cue get go github.com/perses/perses/pkg/model/api/v1
-
-package v1
-
-_#projectDoesNotExistPrefix: "metadata.project \""
-_#projectDoesNotExistSuffix: "\" doesn't exist"
