--- pkg/model/api/v1/project_error.go
@@ -1,33 +1,0 @@
-// Copyright The Perses Authors
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package v1
-
-import (
-	"regexp"
-)
-
-const (
-	projectDoesNotExistPrefix = "metadata.project \""
-	projectDoesNotExistSuffix = "\" doesn't exist"
-)
-
-var projectDoesNotExistPattern = regexp.MustCompile(regexp.QuoteMeta(projectDoesNotExistPrefix) + `.*` + regexp.QuoteMeta(projectDoesNotExistSuffix) + `$`)
-
-func ProjectDoesNotExistErrorMessage(projectName string) string {
-	return projectDoesNotExistPrefix + projectName + projectDoesNotExistSuffix
-}
-
-func IsProjectDoesNotExistErrorMessage(message string) bool {
-	return projectDoesNotExistPattern.MatchString(message)
-}
